+++ app/src/main/java/com/balch/auctionbrowser/MainActivity.kt
@@ -37,6 +37,7 @@ import android.view.View
 import android.widget.SearchView
 import com.balch.auctionbrowser.R.id.*
 import com.balch.auctionbrowser.R.menu.options_menu
+import com.balch.auctionbrowser.auction.AuctionPresenter
 import com.balch.auctionbrowser.auction.model.EBayModel
 import com.balch.auctionbrowser.base.PresenterActivity
