+++ app/src/main/java/com/balch/auctionbrowser/auction/AuctionModule.kt
@@ -20,10 +20,10 @@
  *
  */
 
-package com.balch.auctionbrowser
+package com.balch.auctionbrowser.auction
 
 import android.arch.lifecycle.ViewModelProviders
-import com.balch.auctionbrowser.auction.AuctionView
+import com.balch.auctionbrowser.MainActivity
 import com.balch.auctionbrowser.dagger.ActivityScope
 import dagger.Module
 import dagger.Provides
