+++ app/src/main/java/com/balch/auctionbrowser/auction/AuctionViewModelFactory.kt
@@ -20,11 +20,10 @@
  *
  */
 
-package com.balch.auctionbrowser
+package com.balch.auctionbrowser.auction
 
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
-import com.balch.auctionbrowser.auction.AuctionAdapter
 import com.balch.auctionbrowser.auction.model.EBayModel
 import com.balch.auctionbrowser.note.NotesModel
 import javax.inject.Inject
