+++ app/src/main/java/com/balch/auctionbrowser/auction/model/EBayApi.kt
@@ -22,7 +22,6 @@
 
 package com.balch.auctionbrowser.auction.model
 
-import com.balch.auctionbrowser.AuctionData
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
