+++ app/src/main/java/com/balch/auctionbrowser/dagger/ActivityBindingModule.kt
@@ -22,7 +22,7 @@
 
 package com.balch.auctionbrowser.dagger
 
-import com.balch.auctionbrowser.AuctionModule
+import com.balch.auctionbrowser.auction.AuctionModule
 import com.balch.auctionbrowser.MainActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
