+++ app/src/main/res/layout-land/dialog_auction_detai.xml
