+++ app/src/main/res/layout/dialog_auction_detai.xml
