+++ app/src/main/res/layout/item_auction_list.xml
