+++ app/src/main/res/layout/view_auction.xml
