+++ app/src/main/res/layout/view_label_text.xml
